+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Only write entries that are worth mentioning to users.
 - UI: Remove unnecessary trailing spaces in console output
 - Core: Throw error when there are unsupported message parts
 - MetaCmd: Add `/yolo` meta command to enable YOLO mode after startup
+- Tool: Add approval request for MCP tools
 
 ## [0.52] - 2025-11-10